[6843] - extract PersonalDetailsForm validation
