+++ app/forms/personal_details_form.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class PersonalDetailsForm < TraineeForm
-  include DatesHelper
+  include PersonalDetailsValidations
 
   FIELDS = %i[
     first_names
@@ -25,12 +25,6 @@ class PersonalDetailsForm < TraineeForm
 
   before_validation :set_nationalities_from_raw_values
 
-  validates :first_names, presence: true, length: { maximum: 50 }
-  validates :last_name, presence: true, length: { maximum: 50 }
-  validates :middle_names, length: { maximum: 50 }, allow_nil: true
-  validates :date_of_birth, presence: true
-  validate :date_of_birth_valid
-  validates :sex, presence: true, inclusion: { in: Trainee.sexes.keys }
   validates :other_nationality1,
             :other_nationality2,
             :other_nationality3,
